Typo on Poimandres
I'm maintaining react-use-gesture. Happy to see this lib grow on Vue but I'd appreciate if you mention right away that this is a port of the lib, on top of the Read Me. I'd also appreciate if you could use a different color scheme for the documentation website. This is not urgent, take your time, and if you don't want to, so be it ;)